requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord.py==1.0.0
idna==2.8
multidict==4.5.2
websockets==6.0
yarl==1.3.0
